+++ pastis/e2e_simulators/webbpsf_imaging.py
@@ -10,7 +10,7 @@ import logging
 import poppy
 
 from pastis.config import CONFIG_PASTIS
-import pastis.util_pastis as util
+import pastis.util as util
 
 log = logging.getLogger()
