+++ pastis/image_pastis.py
@@ -18,7 +18,7 @@ import poppy
 import hcipy
 
 from pastis.config import CONFIG_PASTIS
-import pastis.util_pastis as util
+import pastis.util as util
 
 log = logging.getLogger()
