+++ pastis/launchers/run_luvoir.py
@@ -8,7 +8,7 @@ from pastis.config import CONFIG_PASTIS
 from pastis.hockeystick_contrast_curve import hockeystick_curve
 from pastis.matrix_building_numerical import num_matrix_multiprocess
 from pastis.pastis_analysis import run_full_pastis_analysis
-import pastis.util_pastis as util
+import pastis.util as util
 
 
 if __name__ == '__main__':
