+++ pastis/matrix_building_numerical.py
@@ -21,7 +21,7 @@ import numpy as np
 import hcipy
 
 from pastis.config import CONFIG_PASTIS
-import pastis.util_pastis as util
+import pastis.util as util
 from pastis.e2e_simulators.hicat_imaging import set_up_hicat
 from pastis.e2e_simulators.luvoir_imaging import LuvoirAPLC
 import pastis.plotting as ppl
