+++ pastis/pastis_analysis.py
@@ -18,7 +18,7 @@ from pastis.e2e_simulators.hicat_imaging import set_up_hicat
 from pastis.e2e_simulators.luvoir_imaging import LuvoirAPLC
 from pastis.matrix_building_numerical import calculate_unaberrated_contrast_and_normalization
 import pastis.plotting as ppl
-import pastis.util_pastis as util
+import pastis.util as util
 
 log = logging.getLogger(__name__)
