+++ pastis/util.py
